use the schlundtech api to determine subdomain and domain
As there are domains with endings with a various amount of dots, a simple
split is ot possible. A function has been added to inquire the xmlgateway
about the correct domian/subdomain split. This is accomplished by chopping
of the leftmost subdomain in turns and sending an inquiry until a valid
response is returned.